src/datapipeline/synthetic_label_generation/pyproject.toml
[project]
name = "synthetic-label-generation"
version = "0.1.0"
description = "Generate synthetic bite descriptions using OpenAI API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]